requirements.txt
fastapi>=0.115,<1.0
uvicorn[standard]>=0.30,<1.0
pandas>=2.2,<3.0
requests>=2.31,<3.0
pydantic>=2.7,<3.0
python-dotenv>=1.0,<2.0
pytest>=8.0,<9.0
httpx>=0.27,<1.0
